requirements.txt
streamlit
openai
pydub
nltk
python-dotenv
ffmpeg-python
yt-dlp
spacy
